+++ lib/src/property_templates.dart
@@ -3,6 +3,7 @@ import 'dart:math' as math;
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
+import 'collections.dart';
 import 'controls.dart';
 import 'editors.dart';
 import 'icons.dart';
@@ -12,6 +13,7 @@ import 'theme.dart';
 
 part 'property_templates/curve_profile.dart';
 part 'property_templates/attribute_search.dart';
+part 'property_templates/object_picker.dart';
 part 'property_templates/layer_selector.dart';
 part 'property_templates/color_management.dart';
 part 'property_templates/curve_profile_model.dart';